+++ core/src/main/java/io/grpc/internal/Http2ClientStreamTransportState.java
@@ -28,7 +28,8 @@ import javax.annotation.Nullable;
 /**
  * Base implementation for client streams using HTTP2 as the transport.
  */
-public abstract class Http2ClientStreamTransportState extends AbstractClientStream.TransportState {
+public abstract class Http2ClientStreamTransportState
+    extends AbstractClientStream.TransportState implements StreamIdHolder {
 
   /**
    * Metadata marshaller for HTTP status lines.
