+++ okhttp/src/main/java/io/grpc/okhttp/OutboundFlowController.java
@@ -231,7 +231,7 @@ class OutboundFlowController {
     }
 
     OutboundFlowState(OkHttpClientStream stream) {
-      this(stream.id());
+      this(stream.transportState().id());
       this.stream = stream;
     }
